Refactor cobmat view
